fix excessive cpu usage of pulseaudio-tail.sh
closes #99

Polybar now uses "tail" indicating that the script module continuously
outputs instead of it needing to be called on an interval.

pulseaudio-tail listen function now:

- listen function outputs volume every 0.1 seconds, up from 5 seconds,
for faster response time when volume changes

- listen function now waits for sleep to finish before continuing it's
loop. This was the cause of high cpu usage, and explained why it was
worse if left on for a long time.

Polybar now calls pulseaudio-tail instead of calling volume directly
every second. it now calls pulseaudio-tail and expects the script to
continuously output.